[fix] Fix the response format of check-subsidy-agent and the data format of storage(#45).
Since the subsidy information acquisition flow was divided into existence confirmation and detailed confirmation, it was necessary to revise the response format of the subsidy agent to focus on existence confirmation. In addition, due to specification requirements, it was necessary to change the data storage format from json to csv.